requirements.txt
torch>=1.8.0
transformers>=4.0.0
peft>=0.1.0
numpy>=1.18.0
pandas>=1.0.0
scikit-learn>=0.24.0
imbalanced-learn>=0.7.0
matplotlib>=3.0.0
plotly>=4.0.0
streamlit>=1.0.0
python-dotenv>=0.15.0
datasets>=1.4.0
pytest>=6.0.0
